Added vertical layout of for pages.
Updated ReaderViewController to be able to layout pages in a vertical
manner instead of just horizontal as it is now.
To enable vertical layout, in ReaderConstants.m set
READER_VERTICAL_LAYOUT TRUE and run project. Default value for
READER_VERTICAL_LAYOUT is FALSE and so horizontal layout will be
default.